requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.5.2
boto3==1.6.19
botocore==1.9.23
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
docutils==0.20.1
exceptiongroup==1.3.0
fastapi==0.121.0
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.16.0
idna==3.11
jmespath==0.10.0
protobuf==5.29.5
PyAudio==0.2.14
pydantic==2.10.6
pydantic-core==2.27.2
python-dateutil==2.6.1
python-dotenv==1.0.1
requests==2.32.4
s3transfer==0.1.13
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
typing-extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
websockets==13.1